listing_folder_benchmarks/requirements.txt
PyYAML
# Optional for plotting in your local environment:
# matplotlib
# pandas
